Two wires can now merge, just need an alogrithm to detect when wires need to merge

+++ public/components/wire.js
@@ -7,9 +7,10 @@ function roundCoords(x, y) {
 class WireManager {
     constructor() {
         //This will store the node that is currently being drawn
-        this.wireIdGenerator = 1;
+        this.wireIdGenerator = 0;
         this.currentlyDrawing = undefined;
-        this.wires = {}; // Key : (x,y) --> Value: representative node
+        this.wires = {}; //key : id --> Value : line representing the wire 
+        this.wireGrid = {}; // Key : (x,y) --> Value: representative node
         this.handleMouseMoveBound = this.handleMouseMove.bind(this);
         this.handleMouseUpBound = this.handleMouseUp.bind(this);
         this.drawing = false;
@@ -17,17 +18,20 @@ class WireManager {
     }
 
     getConnections() {
-        this.mergeWires();
-//        for (let i = 0; i < wires.length; i++) {
-//            wires[i].connectednodes = [];
-//            this.mapwireline(collection[i]);
-//        }
+        //Hard resets all connections and the wiregrid stored by the wire manager, 
+            //and recalculates everything.
+            //if application gets too slow its probably this function
+        this.wireGrid = {};
+        for(const key in this.wires){
+            const wire = this.wires[key]; 
+            wire.connectedNodes = [];
+            this.mapWireLine(wire);
+        }
+        
     }
     Start(node1) {
         this.drawing = true;
         this.InitialiseWire(node1.getPos());
-//        this.getConnections();
-//       this.mergeWires();
     }
 
     InitialiseWire(start){
@@ -35,11 +39,11 @@ class WireManager {
         //Based on the users mouse movements
         this.defined = false;
 
-        const nHWire = new hline(start,start,this,false,this.wireIdGenerator++);
-        this.wires[this.wireIdGenerator] = nHWire;
+        const nHWire = new hline(start,start,false,++this.wireIdGenerator,this);
+        this.wires[nHWire.id] = nHWire;
 
-        const nVWire = new vline(start,start,this,false,this.wireIdGenerator++);
-        this.wires[this.wireIdGenerator] = nVWire;
+        const nVWire = new vline(start,start,false,++this.wireIdGenerator, this);
+        this.wires[nVWire.id] = nVWire;
 
         nHWire.addEndPointWire(nVWire);
         nVWire.addEndPointWire(nHWire);
@@ -56,7 +60,10 @@ class WireManager {
         const hline = this.wires[this.wireIdGenerator-1];
         const vline = this.wires[this.wireIdGenerator];
         const start = hline.start;
-
+        
+        if(!(hline && vline)){
+            return;
+        }
         //If the user has not moved sufficiently far enough to determine the shape of the wire
         if(this.defined == false){
             if(end[0] !== start[0]){
@@ -76,61 +83,50 @@ class WireManager {
     handleMouseUp(event){
         document.removeEventListener("mouseup",this.handleMouseUpBound);
         document.removeEventListener("mousemove",this.handleMouseMoveBound);
-
-        const end = roundCoords(event.clientX, event.clientY);
-        const start = this.wires[this.wireIdGenerator].start;
        
+
+        const end = roundCoords(event.clientX, event.clientY);  
         const hline = this.wires[this.wireIdGenerator-1];
         const vline = this.wires[this.wireIdGenerator];
-
-        if(start[0] === end[0]){
-            //Delete the wire in the case that the x coordinate did not vary from the start
-            hline.delete();
-            delete this[this.wireIdGenerator-1];
-        }
-        else{
-            hline.updateEnd(end);
-        }
-
-        if(start[1] === end[1]){
-            //Delete the wire in the case that the y coordinate did not change
-            vline.delete();
-            delete this.wires[this.wireIdGenerator];
-        }
-        else{
-            vline.updateEnd(end);
-        }
+        const start = hline.start || vline.start;
         
-//        vline && this.mapWireLine(vline.line);
-//        hline && this.mapWireLine(hline.line);
+        if(hline && start[0] === end[0]){
+            hline.delete();
+        }
+        if(vline && start[1] === end[1]){
+            vline.delete();
+        }
     }
 
     mapWireLine(wire) {
-        const pos = wire.getBoundingClientRect();
+        const pos = wire.line.getBoundingClientRect();
         // console.log(wire);
 
         for (let i = pos.x; i <= pos.x + pos.width - pos.height; i += 20) {
-            if (!([i, pos.y] in this.connections)) {
-                this.connections[[i, pos.y]] = wire;
+            if (!([i, pos.y] in this.wireGrid)) {
+                this.wireGrid[[i, pos.y]] = wire;
             } else {
-                // console.log(this.connections[i,pos.y])
-                wire.connectedNodes = this.connections[[i, pos.y]].connectedNodes;
+                // console.log(this.wireGrid[i,pos.y])
+                wire.connectedNodes = this.wireGrid[[i, pos.y]].connectedNodes;
             }
         }
 
         for (let i = pos.y; i <= pos.y + pos.height - pos.width; i += 20) {
-            if (!([pos.x, i] in this.connections)) {
-                this.connections[[pos.x, i]] = wire;
+            if (!([pos.x, i] in this.wireGrid)) {
+                this.wireGrid[[pos.x, i]] = wire;
             } else {
-                wire.connectedNodes = this.connections[[pos.x, i]].connectedNodes;
+                wire.connectedNodes = this.wireGrid[[pos.x, i]].connectedNodes;
             }
         }
     }
 
-    mergeWires() {
-
+    mergeWires(wire1, wire2) {
+        wire1.merge(wire2);        
     }
 
+    deleteWire(wireID){
+        wireID in this.wires && delete this.wires[wireID];
+    }
 }
 
 export default WireManager;
